feat (home.component.html, home.component.scss, home.component.ts, profile.component.html, profile.component.ts, avatar.lib.ts): implements user avatar display
Adds a user avatar display feature using the dicebear library. This feature generates avatars based on the user's name and displays them in the home and profile components. The changes include adding a new avatar.lib.ts file, modifying the home and profile component templates to display the avatar, and updating the corresponding TypeScript files to generate and assign the avatar URL. Also sets a min-width to the user trigger element.